+++ python/ray/rllib/agents/sac/trainer.py
@@ -1,5 +1,5 @@
 from ray.rllib.agents.sac.config import DEFAULT_CONFIG
-from ray.rllib.agents.sac.sac.sac_policy import SACTFPolicy
+from ray.rllib.agents.sac.sac_policy_graph import SACTFPolicy
 from ray.rllib.agents.sac.proxy.agents.dqn.trainer import GenericOffPolicyTrainer
 
 SACTrainer = GenericOffPolicyTrainer.with_updates(
